All app classes into main
Not changing sys.path at runtime anymore
